Add the ability to add users to a watch list
I might change this to utilize SQLite instead of Viper in the future.
Viper was chosen for this iteration because I figure the amount of
users that will be on the watch list at any given time won't be that
high, and it's a bit simpler to manage, plus users can manually edit
the chatterm.json file if they desire.

Also added a Reset method to the autocomplete to clean up the reset
logic and fix some small issues it had.