BREAKING CHANGE: Store Unix timestamp on messages
Previously just the HH:MM format was being stored in the DB. This meant
we lose the date that the message was received. These changes make it so
we store the Unix timestamp instead, allowing us to use the date as
well.
The /info command will now also group all messages by date.
If you have an existing DB before this commit, you'll probably have to
either delete the content of the chat_messages table via SQLite
directly, or just delete the chatterm.db file in the .config folder.